Still working on expressions/queries. Not running currently but I think I have a good idea of how I need to organise it and some reduction/merging/rationalizing of classes

+++ ArtefactService/Queryable.cs
@@ -262,7 +262,7 @@ namespace Artefacts.Service
 		/// <returns>A hash code for this instance that is suitable for use in hashing algorithms and data structures such as a hash table.</returns>
 		public override int GetHashCode()
 		{
-			return Id;
+			return (int)Id;
 		}
 
 		/// <summary>
